The order functionality is fully operational, and the update feature is working, though it may need some review for improvements

+++ routes/shippers_order.py
@@ -14,7 +14,26 @@ from utils.validate_data import validate_date
 
 order_routes = Blueprint('order_routes',__name__)
 
-@order_routes.route('/orders', methods=['POST'])
+@order_routes.route('/shippers/orders')
+
+# Currently, the function displays all orders because we don't have a shipperId or customerId to filter and show only the orders associated with a specific shipper. Eventually, this ID should be retrieved from the verifyToken middleware to enable proper filtering
+
+def index():
+  try:
+    orders = session.query(Order).all()
+
+    if not orders:
+      return jsonify({'error': 'Order not found'}), 404
+    
+    order_list = [order.to_dict() for order in orders ]
+    return jsonify(order_list),200
+  
+  except Exception as e: 
+    return jsonify({'error': str(e)}), 400
+
+@order_routes.route('/shippers/orders', methods=['POST'])
+
+# This function requires a shipperId or customerId to create a new order. For the time being, I manually pass the required ID, but ideally, it should be retrieved through the verifyToken middleware.
 
 def create_order():
   try:      
@@ -24,13 +43,16 @@ def create_order():
 
     if validate_result.get('error'):
       return jsonify(validate_result.get('message')), 400
+    if order_data.get('orderStatus') != 'pending':
+        return jsonify({'error': 'Invalid status for shipper'}), 400
+    
     else:
       new_order = Order(
         customerId = order_data.get('customerId'),
         driverId = order_data.get('driverId'),
         pickupLocation = order_data.get('pickupLocation'),
         dropoffLocation = order_data.get('dropoffLocation'),
-        orderStatus = order_data.get('orderStatus'),
+        orderStatus = 'pending',
         paymentAmount = order_data.get('paymentAmount'),
         vehicleType = order_data.get('vehicleType'),
         dimensions = order_data.get('dimensions'),
@@ -39,12 +61,25 @@ def create_order():
       )
       session.add(new_order)
       session.commit()
-      return jsonify(order_data), 201
+      return jsonify(new_order.to_dict()), 201
     
   except Exception as e:
     return jsonify({'error': str(e)}), 400
   
-@order_routes.route('/orders/<id>', methods=['PUT'])
+@order_routes.route('/shippers/orders/<id>', methods=['GET'])
+
+def show(id):
+  try:
+    order = session.query(Order).filter_by(orderId=id).first()
+
+    if not order:
+      return jsonify({'error': 'Order not found'}), 404
+    
+    return jsonify(order.to_dict())
+  except Exception as e: 
+    return jsonify({'error': str(e)}), 400
+
+@order_routes.route('/shippers/orders/<id>', methods=['PUT'])
 
 def update(id):
   try:  
@@ -59,46 +94,29 @@ def update(id):
     if validate_result.get('error'):
       return jsonify(validate_result.get('message')), 400
     else:
-      order.customerId = order_data.get('customerId')
-      order.driverId = order_data.get('driverId')
       order.pickupLocation = order_data.get('pickupLocation')
       order.dropoffLocation = order_data.get('dropoffLocation')
-      order.orderStatus = order_data.get('orderStatus')
-      order.paymentAmount = order_data.get('paymentAmount')
       order.vehicleType = order_data.get('vehicleType') 
       order.dimensions = order_data.get('dimensions')
       order.weightValue = order_data.get('weightValue')
-      order.deliveryTime = order_data.get('deliveryTime')
     session.commit()
 
-    order_dict = {
-            'orderId': order.orderId,
-            'customerId': order.customerId,
-            'driverId': order.driverId,
-            'pickupLocation': order.pickupLocation,
-            'dropoffLocation': order.dropoffLocation,
-            'orderStatus': order.orderStatus.value,
-            'paymentAmount': order.paymentAmount,
-            'vehicleType': order.vehicleType.value,
-            'dimensions': order.dimensions,
-            'weightValue': order.weightValue,
-            'deliveryTime': order.deliveryTime
-        }
-    return jsonify(order_dict), 200  
+    return jsonify(order.to_dict()), 200  
       
   except Exception as e:
     return jsonify({'error': str(e)}), 400
   
-@order_routes.route('/orders/<id>', methods=['DELETE'])
+@order_routes.route('/shippers/orders/<id>', methods=['DELETE'])
 
 def delete(id): 
   try:  
     order = session.query(Order).filter_by(orderId=id).first()
-    
+
     if not order:
         return jsonify({'error': 'Order not found'}), 404
-    session.delete(order)
-    session.commit()
+    if order.orderStatus != 'on-route':
+      session.delete(order)
+      session.commit()
 
     return jsonify("Order deleted successfully"),200
   except Exception as e:
